test(skills): split specialist procedure coverage
Separate contract closure, validation mutations, reviewer outputs, and generated parity into focused test modules with shared fixture support. Preserve all 11 original test names and assertion behavior without touching production or generated files.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: Preserve every specialist procedure contract and generated-byte assertion; change test organization only.
Rejected: SIZE_OK exception | contract, review, validation, and parity responsibilities split cleanly below 250 pure LOC.
Confidence: high
Scope-risk: narrow
Directive: Keep future specialist procedure tests separated by validation, reviewer contract, and generated parity.
Tested: 26 focused specialist tests; 11-test name and assertion equality; Ruff; compileall; pure-LOC gate; git diff --check.
Not-tested: Full discovery and remote CI after this test-only split.
Plan: .omo/plans/omh-domain-skill-quality-upgrades.md
